move `cmd/mdatagen` from internal/core
