+++ cmd/mdatagen/README.md
