+++ cmd/mdatagen/main.go
