+++ cmd/mdatagen/metrics.tmpl
